+++ packages/react-native-aria/menu/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @react-native-aria/menu
 
+## 0.2.12
+
+### Patch Changes
+
+- fix: warnings of deprecated props
+- Updated dependencies
+  - @react-native-aria/utils@0.2.11
+  - @react-native-aria/interactions@0.2.13
+  - @react-native-aria/overlays@0.3.12
+
 ## 0.2.11
 
 ### Patch Changes